Support the circle pad pro and New 3DS c-stick
Still needs some cleanup and testing on Windows, but now the CPP and
New 3DS extra inputs should work.

Also switched the network communication to send the full input state
instead of individual changes. This simplifies the code and enables us
to (in the future) switch off of TCP (to improve latency).

This uses some ctru-rs changes that are still in PR:
https://github.com/rust3ds/ctru-rs/pull/86

The C-stick calibration happens automatically, though there might be a
better way to do this (maybe client side?).